Fix rewriting scales
If both the `entry_teaser` and the `listing_thumb` images are missing,
creating them rewrites all the content of 'plone.scale' in the
__annotations__.
For this reason when `entry_teaser` is created it has a valid URL that
will turn into a 404 as soon as the `listing_thumb` scale is created.

The patch avoids that and also simplifies the tal:condition statements.